third_party/WebKit/LayoutTests/web-animations-api/player-finish-repeats-expected.txt

PASS: playerTop is target
PASS: event currentTime equals player currentTime
PASS: event timelineTime equals timeline currentTime
PASS: playerTop is target
PASS: event currentTime equals player currentTime
PASS: event timelineTime equals timeline currentTime
PASS: playerMiddle is target
PASS: event currentTime equals player currentTime
PASS: event timelineTime equals timeline currentTime
PASS: playerMiddle is target
PASS: event currentTime equals player currentTime
PASS: event timelineTime equals timeline currentTime
PASS: playerBottom is target
PASS: event currentTime equals player currentTime
PASS: event timelineTime equals timeline currentTime
PASS: playerBottom is target
PASS: event currentTime equals player currentTime
PASS: event timelineTime equals timeline currentTime
